neural_style/requirements.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
numpy>=1.23.0
Pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
matplotlib>=3.5.0
python-multipart>=0.0.5
streamlit>=1.30.0
